+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Fixed
+- Fixed issue where parent records were not getting deleted from solr if empty. (DR-2342)
+
 ### Updated
 - Added additional rights statements that release high resolution permalinks (DR-2396)